Parallel runs for the off-site spin-orbit implementation.
- Avoided the use of transpose matrices.

- Now parallel and serial runs can be performed when the 
off-site SO is included in the calculation. 

- Tested the former on-site spin-orbit implementation.
It works as in the previous SIESTA versions. However, 
the off-site spin-orbit formalism has set as the default 
fully relativistic option. If the on-site wants to be 
used one has to change explicitly the corresponding 
flag in the fdf file (See manual for further info). 

- LDA+U has to be checked as well. So far it is not tested
both features.


modified:
  Src/Makefile
  Src/SiestaXC/ggaxc.f
  Src/SiestaXC/makefile
  Src/atmfuncs.f
  Src/compute_dm.F
  Src/compute_energies.F90
  Src/dfscf.f
  Src/dhscf.F
  Src/diag3g.F
  Src/diag3k.F
  Src/m_energies.F90
  Src/m_new_dm.F90
  Src/nlefsm.f
  Src/rhoofd.F90
  Src/scfconvergence_test.F
  Src/setup_H0.F
  Src/setup_hamiltonian.F
  Src/siesta_forces.F90
  Src/siesta_options.F90
  Src/sparse_matrices.F
  Src/state_init.F
  version.info