Updated tests folder, added FLOS tests for LBFGS, FIRE and meshcutoff
Added flos tests.
To run the flos tests it is required that:
1. SIESTA has been compiled with flook support
2. flos has been installed by adding the library to the LUA_PATH
   environment variable.

The new flos tests involve a:
 LBFGS relaxation with the H2O molecule
 FIRE relaxation with the H2O molecule
 MeshCutofff convergence with the H2O molecule
 NEB of rotation of H2O molecule in a two molecule setup (not enough vacuum
 , but useful enough for testing).

all 3 are very fast and shows how a Lua implementation could
be performed.

This also required that the tests folder is updated.
Now the tests are subdivided into specific tasks:
 tests_soc; all spin-orbit tests. These are typically
 	   very loong tests and require a long time to run.
 tests_flos; tests requiring flook and flos
 tests; everything else

Later these may be extended.