Fixed an out-standing bug related to bond-currents and dH (TBT)
The delta-H method in TBtrans has been revised and updated.

Now dH adds e ^ { i k.R } phases for consistency. If people use it
for rescaling coupling elements this makes sense, if it is used
for more advanced uses (where the phases shouldn't be there) the
users should be familiar with this and add the opposite phases.

Secondly, the bond-currents are now defaulting to adding the dH
elements to the bond-currents. I.e. the same equation for H
is now also applied with dH. This is controlled via the flag

  TBT.dH.Current.Orb

which defaults to True when bond-currents are requested.

The documentation has been updated to reflect the changes.