Enabled Lua calls with the new mixing routines
The Lua-interface to changing the mixing weight etc. was broken
because of the new mixing routines added.
This has now been re-enabled because it seems like a very useful feature.

A couple of requirements were needed to be fulfilled:

1. The fdict variable required an update to enable pointers
   to character arrays (of len=1).

2. The ncdf library then also required an update.

3. The internals of the character retrieving functions needed adaption
   of the new method in fdict.

Now this is much better and easier to handle.

Two examples are now for the "pure" lua interface (not flos):

- lua_h2o
  An adaptive change of the mixing weight per iteration.
  This is in an MD environment and thus it resets it every
  MD step.
  
- lua_si111
  A mixing method changer based on the convergence criteria.
  This will change the mixer to more aggressive methods at
  certain convergence levels.

Added a example explanations in the documentation.