Provide units for max-walltime and max-walltime-slack. Add fdf_isphysical
The fdf options max-walltime and max-walltime-slack should properly be
specified with units (e.g. 2.5 hours), instead of what a single number
interpreted as seconds.

To maintain compatibility with previous usage, a new boolean function

 fdf_isphysical(label)

has been added to the FDF library, to check whether units are given for
the value of 'label' in the fdf file or not.