Updated all tests for the 4.0.1 release
Every test has been updated and checked.

There were some tests were the energy was changed.
This was due to fixes in the dipole correction where the
energy contribution of an internal field is only half of
an external field. This was a bug in prior versions.

Secondly, there were a few tests missing in the Tests/Makefile
which have now been added.