Enabling the memoryfactor option
It is unclear whether it is actually needed, however,
in case something goes wrong it may be nice to have.

Updated the work array sizes for the lapack and scalapack routines
where the work-query does not provide details.

All different algorithms have been runned on the si001 test.
Note I haven't tested all the real algorithms.

Added a bug-printout if compiled with _DIAG_WORK.
In this case if ScaLAPACK routine pzheev does not yield the
correct work-size, a note will be written that there is a bug
in ScaLAPACK.

At least in 2.0.2 there is a bug (I have reported a bug, see
http://icl.cs.utk.edu/lapack-forum/viewtopic.php?f=13&t=5164)