Enabled NumberOfEigenstates for all diag* routines
Now all the k-point routines uses the NumberOfEigenstates
when calculating the initial eigen-spectrum, but subsequently
it is adjusted to the minimum # of eigenstates.
Note that there should be no problems by enabling this for the
k-point sampling, if it "goes" wrong it has to be the users
responsibility.

Language corrections in the manual, it is written in US English,
while diagonalisation was written in British English.

Allowed the new diag routines to have 3 neig clauses:

 neig > 0:
   only calculate neig eigenvalues AND eigenvectors
 neig < 0:
   only calculate neig eigenvalues
 neig == 0:
   calculate all eigenvalues

Tested the non-collinear and the spin-orbit cases.