Bug fix for NumberOfEigenstates, enabled new option
The selection of the number of eigenstates was buggy for
non-polarized non-collinear and spin orbit cases, however,
it was only if the numberofeigenstates option was too small.
This should have been corrected now.

The option NumberOfEigenstates now implements a variant of specification
that should make it easier to select the "correct" number of eigenstates.
If one passes a negative number it corresponds to the number of
calculated eigenstates above the number of orbitals required to fill the
orbitals according to the charge.
I.e. if there are 20 orbitals with 10 electrons and -2 is passed,
it would be equivalent to NumberOfEigenstates 12.

Added note about possible buggy pdsyevr when requesting less than the number
of total eigenvectors.