Added ELPA-2017, however it does not work :(
I have now added ELPA to the diag.F90 code.
It should work, as far as I can see. But there *must* be something
wrong.

The implementation of the ELPA library requires the current
implementation of ELPA (>=2017.05).

Added index to the solvers in the documentation.
The documentation has also been updated with ELPA, but
it is currently not shown (easy to enable once it works).

A fix for possible problems with too few processors per column
has been fixed.

Added neig to all the ScaLAPACK routines for back-transformation and
the redistribution routines to limit computations and communication
overhead.

TODO: re-run all tests with NumberOfEigenstates to assert the reduced
computations.