Enabled a list-like test for the makefile system
To easily enable testing a variety of choices of fdf
variables one may now create a single test makefile
which defines an fdf-name, and a list.
I.e. an example makefile in a test folder may look
like this:

   FDF_LIST = Diag.Algorithm
   LIST = divide-and-conquer expert noexpert mrrr

   include ../test-list.mk
   name=si001-diags

which will then (when executing make) in the folder
run all the algorithm variables in separate work_* folders.
This may come in handy to check several values, but otherwise
retaining the same test structure.

See si001-diags for an example use.