Enabled 1D distribution with ELPA and made sure it works
I found out that the current ELPA implementation
actually works. But only IFF I use a 2D distribution
and use all processors on the row (so grid-column == 1).
This is rather unfortunate, but is probably due to a bug
in ELPA. For now, if one requests ELPA, we force the above.

Hence, I believe it to be a bug in ELPA that it does not
work for the other distributions. I will let ELPA devs know.

Fixed a bug for expert driver in non-2D distribution.

Changed the behaviour of 2D grids. Now it only makes a
2D grid IFF it is able to split in more than 1 dimension.
I.e. Y x Z for Y > 1 and Z > 1. Otherwise it defaults to
1D distribution (but not for ELPA yet).
I.e. if the 2D distribution would be an Y x 1 or 1 x Y it makes
no sense to redistribute data for the same distribution.
However, the user may override these settings.

Updated the documentation to reflect the ELPA usage.

Fixed compilation in serial.

I have now checked all cases of runs and it should have converged.

MRRR is now the default diagonalization routine, if enabled.

Cleanup of code in diag.F90