Enabled a diagonalization blocksize to explicitly set it
The blocksize only gets set for 2D distributions because
we have to redistribute it there.
Hence, if the blocksize is different from the internal blocksize
the 2D distribution is default to .true., regardless of processory.
This may help certain systems where diagonalization is of highest importance.

Updated the documentation slightly.

Fixed print-out if non-compiled option is present (MRRR and ELPA).