Small bug-fixes and corrections from before this branch
Added print-out and quick-return when requesting band wavefunctions
for spin-orbit coupling.

Added note about neigneeded in diag*k being explicitly calculated and
hence no 2* is needed for non-colinear or spin-orbit.

Fallback the block-size when only using 1D distribution in diagonalization
routines. This is merely cosmetic to let the user know the used block-size.

Fixed an neigwanted in an rdiag call when it fails the diagonalization.

Added comments for why we can re-use the 1D matrices in a 2D distribution
when the memory requirements are equal or smaller in the 2D case.