Updated some tests, partially fixes lp:1712317
Added makefiles for:
  Util/Contour
  Util/Optical
  Util/sies2arc

to enable easier compilation of said utilities.

Removed clean targets in the protoneb and simpletests in
siesta as subroutine utility.

Readme files in above directories are also updated.

In Optical a program called "input" has been changed to
optical_input to prevent name clashes.

I haven't implemented the LINALG fix, but I think we should
encourage the use of external libraries for performance reasons.
Essentially the shipped sources for LA are highly *discouraged*.