Added total energy contribution from an open-boundary calculation.
The total energy contribution is (grand canonical ensemble):
  Etot = Etot - e \sum_i N_i \mu_i
where N_i is the number of particles being occupied from the i'th chemical
potential, with chemical potential \mu_i.
Sadly this value depends on the choice of reference energy.
For choices:
  mu_1 = -V/2
  mu_2 = V/2
and 
  mu_1 = 0
  mu_2 = V
one will find two _very_ different results. But this is a matter of choosing the
zero potential.
Perhaps we should make the "effective" chemical potentials be subtracted the
average chemical potential to have a constant potential for varying bias'.