Changed internal Gamma array to be i (Sigma - Sigma^\dagger)^T
Generically the scattering matrix (Gamma) is defined as:
  Gamma = i ( Sigma - Sigma^\dagger)
however, before this commit the convention was:
  Gamma = ( Sigma - Sigma^\dagger) ^ T

Now we use this convention:
  Gamma = i ( Sigma - Sigma^\dagger) ^ T

1. The complex number i makes the scattering matrix Hermitian
2. The transpose allows certain optimizations when calculating the
   transport (i.e. dot instead of matrix multiplications)

I have re-runned all k-point sampled versions of MUMPS, full, BTD
with and without bias (only bias should be important) and I have
also checked TBtrans.
The only thing missing for testing is the molecular projections
for tbtrans. This should be tested (TODO). However, there it shouldn't
matter because the Gamma are not re-calculated.

I also found a bug when writing correction eigenvalues to ASCII,
now it correctly only writes the k-averaged ones in AVCEIG.