Updated dH method to be a dH and dSE
This is necessary to differentiate between delta elements contributing to
the bond-currents (magnetic field) and which does not (complex absorbing potentials).

This has meant a restructuring of the delta stuff and is now easier to follow,
the dH/dSE codes are now extremely minimal due to a common interface.

I have tested them against level 1-4 for both cases.
Enabled print-outs of delta stuff for level > 7.

Updated mat_invert MPI_IN_PLACE_RECURSIVE algorithm to reduce copies, it is still
slower, but is good to have.