Fixes for running TDDFT branch on my box, reduced memory requirement
I couldn't get TDDFT branch to work on my box, I required to rework a few
things to make it work.

In the process I found that one could easily reduce memory requirements
further. Hence this commit further reduces the memory requirements
in the inversion algorithm, so now it is 50% of what it was in the initial
branch.