Fixed pivoting in tbtrans
For a very long time TBtrans has experienced sub-optimal pivoting
compared to transiesta.
This has now been fixed because the (dense) down-folded self-energies were
not expanded in the sparse pattern that is searched when pivoting.
This commit uses the correct sparse pattern.

Secondly, this commit introduces a change to the PCG algorithm which
in certain cases proves to be really efficient. It is slightly more
costly than the GPS algorithm but may provide better pivoting.
It has at least for some really big TB systems shown far superior handling.