Ensured that siesta information gets printed before fdf-stuff
This has historically always been the case and it also makes it
easier to debug.

I have made a slight re-arrangement of code to implement this.
The one downside is that the Nodes print-out will be there
twice in case of PEXSI calls. This I think is an acceptable
case.