Allow passive species in SOC calculations
In previous versions of the code all species had to contribute to the
spin-orbit part of the Hamiltonian. This has been relaxed, and there
is code to check that at least one species contributes.

In the output file there is a new section on SOC setup. The old line
with the confusing message "spin-orbit components not used" has been
removed.