Put libs config step in arch.make. Consolidate MPI and NetCDF operation
Separate the arch.make contents in two sections: user options (including
optional WITH_MPI and WITH_NETCDF variables) and a fixed, slowly varying,
variable-setting section.

In the variable-setting section, process the .mk files for external libraries.

(The logic assumes that libGridXC offers separate serial and mpi versions.
This is yet to be implemented, but a proof-of-concept has been demonstrated
with a synthetic GRIDXC_ROOT setup.)