Fixed a_dev in TBT.SE files (sorted) and also fixed issue with MPI-IO
Now the TBT.Proj and TBT.SE files have their own open routine which is
simply a copy. But it allows a finer tuning.

Fixed warnings when sigma-only and fixed a bug regarding elec-dos and sigma-only.

Speeded up SE writes because we now keep the file-handle open.

Added generalized option for connectivity graphs. We should consider
doing this also for all other options. I.e. the fan-options probably
require a more specific pivoting routine.

Fixed wrong default in the manual of the cross-term updates, they are
cross-terms, but none was specified as the default.