Stored orbitals directly
(So far only for SPLIT and vanilla POLgen)

Had to move the location of the call to atm_pop so that
occupations were computed before the storing of the information.
This results in the movement of a small section of output.

old_atmfuncs not trimmed yet. Some arrays are used for
internal communication within 'atom'.