Reverted an old change in the electrostatics for N-electrode box solutions
In 391.1.642 I changed the algorithm from the in_box to square grids which
meant that the box functionality was *only* working for Cartesian lattice
vectors. This has been reverted.
Secondly the loop has been reduced to only check where the electrode has points.

One of the problems with the old implementation was the extremely bad
performance due to numerous calls of dgesv routine. Instead the m_geom_box
now calculates the inverse of the box-vectors and then multiplies to get
the solution. This makes it perform MUCH faster.

The electrode was using this box method and thus is now much faster.

Removed some die calls in init_Elec_sim and moved to ts_warnings
to make it compatible with N-electrode calculations. Say N-electrode
calculations with Gamma-only don't really care about non-conforming
pivoting of electrode lattice vectors.

Thanks to Thomas Frederiksen for bringing this to my attention.