Speeded up DOS calculations in TBtrans
The calculation of the DOS was performed by looping BTD entries which
is far from optimal when performing large scale calculations.
For now we have reverted to the sparse matrix looping and extraction of
the index from a function call.

This is implemented both in GF_DOS and A_DOS and so far it seems to be twice
as fast. The other benefit is a better OMP parallelization.

Also fixed a compilation problem without NCDF.

Added a new routine in TriMat to extract index in sub-matrix plus submatrix.