Added COOP and COHP for tbtrans
Now tbtrans can calculate COOP and COHP curves. This is stored in much the same spirit as the
orbital-current due to the same sparsity pattern. Indeed this analysis does not make
sense for orthogonal basis. However, there is no check for this.

Basically the COOP curves are calculated as:

   COOP_ij = - Im[Gf_ij * S_ij] / Pi
   COOP_ij = Re[A_ij * S_ij] / 2Pi

and the COHP curves are calculated as

   COHP_ij = - Im[Gf_ij * H_ij] / Pi
   COHP_ij = Re[A_ij * H_ij] / 2Pi

Currently the COOP/COHP curves are untested but are (hopefully) correct.
They are also implemented for dH calculations.

Manual has been updated to reflect these changes.

To make COOP work smoothly I have generalized the cdf_save function of the
orbital current to make it work for all 3 cases (orb-current, COOP, COHP).