Bug re-read basis functions with no potentials to interpolate (fix lp:1736455)
When re-reading a basis from a ghost or Bessel orbital then many of the
quantities are 0 (Vna, VCore, etc.).
Since 4.0 the interpolation routine prevented these from the spline interpolation
because there is a check for monotonic arrays (all 0).

This commit fixes this by simply checking for erroneous input, and if so it will
set dy2/dx2 to 0.
This is rather arbitrary but at least initializes the dy2/dx2 to some values.

Secondly, a possible memory leak in interpolation is corrected by issuing
a fixed clean_spline call.