Bug fixed for COOP and COHP curves in tbtrans
Since they reflect the DOS they should be calculated as Gf_ij S_ji as would
a normal DOS calculation. I.e. the papers describing COOP and COHP curves are only
considering the Gamma-point. I have checked that this is indeed correct by
using sisl and everything in full matrices.

Also ensured that J, COOP and COHP are initialized to 0 which could be important with
dH calculations.

It should be noted that sum(COOP) == total-DOS.