Bug in writing a Bessel function ion/ion.xml file
Since the Bessel basis does not contain any pseudo-potential siesta
would write binary stuff in the ion/ion.xml files which was preventing
it from being read elsewhere.
I have added a pseudo-potential initializer which defaults lots of
values to the pseudo-potential (could be necessary for Bessel and other
exotic pseudos.

Also ensured lots of type arrays to be nullified upon creation, which
should help in the aid when debugging.