Ebs and Eharris bug corrected in compute_energies subroutine.
- Some inconsistencies in the rotational invariance of 
Ebs and Eharris have been corrected for the offsite SO.

- Following the offSite SO DM convention some corrections 
in print_initial_spin and in the user's DM initialization 
in m_new_dm have been modified. The final printed spin in 
Mulliken subroutine has been changed as well. Specifically, 
the sign of Y component of the MM was adapted to OSSO. Maybe 
it has to be added a flag when the OSSO is not used because 
the sign and convention will be different if onsite is used
(even for non-collinear).     


modified:
  Src/compute_energies.F90
  Src/dfscf.f
  Src/m_new_dm.F90
  Src/mulliken.F
  Src/print_spin.F90
  version.info