Cleaned up non-colinear comments and routines
I have now reworked the diagonalization routines for the
non-colinear methods.
It was apparent to me that the correct formulation was diag2k because

  DM(a,b) = <psi_a|psi_b>, a and b orbital and spin indices

which means that the conjugate has to be taken on the LHS eigenstate.
This was opposite in diag2g, however since the imaginary part is only used
in one of the assignments it was not seen.

In effect, the previous codes had no bugs, but they were not consistent
with the formulation.
I have also clarified this in all comments in both routines.

I have checked tests fe_clust_noncollinear and fe_noncol_kp.