Fixed Lua-flook interaction when supercell changes through MD
When the supercell changes, so does the internal variables such
as xa, etc. In that case we *have* to re-establish the pointers
to the dictionary connecting Lua. This should now be fixed in
state_init after the superc call.

When doing this I also found out that several of the arrays where
written as the full arrays (na/o_s) which was never the intent.
Now only na/o_u elements are stored.