Clarified slab-dipole corrections for transiesta calculations
Now the slab-dipole correction seems to be more consistent with the
semi-infinite leads.

1) Any electric field along the semi-infinite directions is not
   allowed.
2) Any dipole correction will be added perpendicular to the semi-infinite
   directions.
   However, for 1 electrode we *only* allow a correction along the single
   semi-infinite direction. This is perhaps not the best method. But it
   resembles that the system indeed is a slab along that direction.

In any case, the dipole correction still needs to be turned on manually.

Also, a long standing dhscf_init bug has finally been fixed.
When running transiesta MD calculations it is important to update coordinates
and indices for the grid. Hence the initializations has to be done
after each MD step.