Introduce MPI_Comm_DFT to clarify hierarchies
When both the MPI-dispatching (fsiesta-mpi) feature and the enlarged
solver MPI set are used, it was not appropriate to use the "true"
world communicator for solver communications. A new communicator,
MPI_Comm_DFT, has been introduced to represent the communications
within a single "DFT calculation" (core Siesta ops + solver).

This is in preparation for the ELSI integration, which will use the
same basic framework as the legacy PEXSI interface. A few more
clarifications will be needed along the way.