Generate timer report in non-convergence and out-of-time conditions
For proper accounting, the timer information should be generated even
if the program stops due to non-convergence (scf, geometry) or
out-of-time conditions.

This is achived by inserting appropriate calls to 'timer' before the
call to 'die'.  To avoid a race condition in which a non-root node
might get to the 'die' call before the IO node has had time to
properly write the CML, timing, and even output information, a call to
'barrier' has been inserted before the 'die' call.

It might be argued that this kind of bookeeping should be left to the
'die' call itself. But we should make sure that all nodes call 'die'.