Huge speed up for very large systems in TBtrans
TBtrans was using a huge amount of time on trying to even-out
the BTD blocks. Now I have reworked this routine and it is *much* faster.

This means that the "speed" method has changed to not start from the
largest values and doing it like that. Searching for the largest element
is extremely slow and is killing performance.
Now there are two routines:

  diff_mem / diff_perf

which compares two BTD matrices and returns a difference number
in terms of their memory/performance, respectively. These
are now used instead and should also work without doing
overflows.

All tbtrans: names has been changed to tbt: TODO consider changing
to specific names depending on their parts.

Enabled associating a region with a list.