Yet another huge speed up for very large systems in TBtrans
TBtrans was using a huge amount of time on trying to create the
connectivity graph due to some of the rgn_sp_* routines.
I have now reworked them so they use slightly more memory, but
are much much faster!
This means that initialization of very big systems are almost
negligeble.