Large speed up in certain pivoting algorithms
Now the sorting of a pivoting array is much faster.
This reduces the overhead of initialization of tbtrans to
a minimum.
I have also removed the constraint that some rgn_sp*
routines enables distributed computing. Not necessary (and
never checked).

Added a rgn_grow routine which only re-initializes a region
if there is not room for the new size.

Allowed rgn_uniq to be in-place (i.e. no re-allocation, simply
reducing r%n to the unique values).