Speeded up pivoting and block finding algorithms, also corrected a memory print-out
I think we are approaching a final stage in the speedup for extreme TB calculations.
The performance is rather good now. Still takes time in the trimat setup because
it "tries" lots of different BTD block-sizes. This should probably be changed such
that it stops guessing as soon as the final block size doesn't match.

Speeded up a region Bcast when Nodes == 1.

valid_tri and full_even_out should still be optimized...