Moved even-out to after valid_tri
This is necessary if the initial BTD is wrong. In that case it makes no sense to
even-out the blocks.

Also reduced the search space by trying to optimize the loop parameters.
Now the initial guess 95% of the min bandwidth within the first connections.
The stepping is 1% of the min bandwidth.
The max is 105% of the max bandwidth within the first connections.