Revert: Leave the XC setting of H(*,7:8) in vmat for now
Even though formally 'vmat' should not have to deal with SO issues, it
is now considered safer to keep the setting of the XC components of
H(*,7:8) in it, rather than doing it explicitly in the H building
routines. 'vmat' is just adding to H, so no overwriting is done. In
contrast, errors would ensue if somebody changed the ordering of the
assignments in 'setup_hamiltonian'.

This patch reverts a previous one in the 4.1-spin-fix branch.