Spin monitoring during scf cycle
If the fdf variable 'Spin.In.Scf' is set to 'true', the size and
components of the (total) spin polarization will be printed at every scf step.
This is analogous to the 'Mulliken.In.Scf' feature.

The initial spin polarization for every geometry iteration is also
printed in 'state_init', replacing the similar calls in 'new_dm'.