Use explicitly the same D12 sign-convention as in SOC versions
In the non-collinear case, the code is actually using internally the
same sign convention for the "up-down" components of H and the DM as
the spin-orbit (SOC) versions: The building of the dense Hamiltonian
has the same form and the off-diagonal components of Vxc are
identical. Routines diag2g and diag2k are actually computing the right
DM, but with DM* as an intermediate step.

By changing the interface to the outside world, namely the
correspondence between spin angles and the DM in new_dm, and the
calculation of the spin (actually moment) components in spnvec, the
code can now interoperate (i.e., reuse DMs) with the SOC versions.