Further control over the output of orbital magnetic moments
For spin-orbit calculations, the orbital magnetic moments are only
printed if the flag WriteOrbMom is set to .true.. Routine 'moments'
has been slimmed down to remove the spin part, already done in
'mulliken', and the moments themselves and not the angular momenta are
printed.

If MullikenInScf is .true., the information is printed at each scf step.