Clarified phase usages in PDOS routines
I have gone through the phases used in the pdos*k routines.
There were a few unclear things that only made sense once
a deeper understanding was taken.

For instance all H(k) are setup with:

  H(i,j) = H(i,j) + e^{-i k.R} * H(ind)

but when performing the matrix-multiplication for the
PDOS calculation <psi_j | S(k) | psi_j> one finds that
S(-k) is setup. This is due to the data-layout in
parallel that requires S to be distributed over rows, and
thus a simpler way to do the matrix multiplication is by
doing element-wise multiplications using S(-k).

In this commit I 1) added comments in pdosk/pdos2k/pdos3k to
clarify this intent.

In pdosk I also changed a few signs to introduce the S(-k)
in the correct place. Before the minus sign was just introduced
at a point that was not so trivially found.

All-in-all. This commit does not change *any* results! I have
runned si_pdos_kgrid with non-polarized, polarized and non-colinear.
(not SO) And all results where the same.

I have also reduced some complexity by using pointers instead of
data-copying. This was actually very simple and cleaned up the
readability quite a bit.