Fix bug and clarify code in dfscf for SOC case
In dfscf there was overcounting of the off-diagonal elements in the SOC case
(first doubling of 3:4 and then copying of 3:4 into 7:8 components).

Besides, the code was confusing, since the grid potential has only four
bona-fide components, but was artificially extended to match the 8-component
DM in the SOC case. A clarification (and simplification) can be achieved if
the DM is explicitly "spin-box hermitified" in the same way as it is done
in the computation of the charge density.