Re-fix known issues kept active for debugging purposes
-- The SO contribution of the highest-l pseudopotential was not
taken into account. This bug had been re-introduced in revno 668.

-- The computation of two-center integrals needs a larger NRTAB for
extra accuracy. This change was not brought over when synching to the
trunk version in revno 668.

The 'npotu' bug is now fixed by default (it is still possible to
re-enable it via the line 'keep-npotu-bug T' in the fdf file).

NRTAB is now hardwired to the trunk value of 1024 (no changes allowed
without editing the code).