Re-implement spin-orbit-strength feature at the atom level
The 'offsite' flavor of SOC does not separate 'ion' and 'SO'
contributions to the forces, so the previous implementation of the SO
factor feature gave wrong forces. It has been re-implemented for now
by modifying the SO parts of the semilocal potentials read from a
\code{.psf} file. Care must be taken when re-using any \code{.ion} files
produced.  Recall that this is only for debugging and experimental
purposes.