Minimal fix for DM initialization with severe folding
When an orbital interacts with its own periodic replicas and a
Gamma-point calculation is requested, the naive DM initialization with
atomic information is faulty: the diagonal S entries are no longer 1.0
(but are enhanced by the contributions from the replicas), so the
normalization of the DM is lost (the code will print a message to the
effect). The initial charge density will be distorted with respect to
its intended form (the result of a blanket normalization). Note that
the program might still go on to converge to the right ground state,
since the code to deal with multiple orbital interactions in the
Gamma-point case is mathematically correct.


There are several levels of fixes for this:

1. Use k-point sampling, as it can be argued that requesting a
Gamma-point calculation for a very small cell is physically
meaningless.

2. Force the use of an auxiliary supercell ('force-aux-cell
T'). Internally, the code will correctly place all the different
interactions in an expanded auxiliary supercell. The drawback is that
complex arithmetic and diagonalization must then be used (there is no
yet support in the code to avoid this).

3. Correct the initialization of the DM so that its norm is conserved
and the initial charge density is more reasonable.

In this patch we implement case 3. above by (always) dividing the
recently-initialized diagonal DM entries by the corresponding S entry.
Note that this is a zeroth-order correction, and the charge density
produced by the initial DM will exhibit a deficit at the orbital
position and 'ringing' enhancements at the superpositions of the
orbital and its periodic images. The initial scf steps will thus show
differences with a calculation done with 'force-aux-cell T', but
eventually they will converge to the same ground state.