Streamline hsparse. Kill the program if auxiliary supercell is inadequate
The 'multiple-image' warnings in hsparse were not representative
enough under MPI operation (only the root node got to report
them). And it can be argued that they are superfluous except in a
'debug' mode of operation.  The key severe diagnostic is the existence
of 'folding' in a non-gamma calculation, which should lead to aborting
the program and to further investigation of the internals of the code.
In the gamma case, folding is reported and a further note is
printed if the user is attempting to carry out a COOP/COHP analysis (which
would be faulty in the presence of folding). 

The non-documented 'debug-folding' feature is implemented very
coarsely for now.