Enabled density matrix calculations in tbtrans
Also updated manual to reflect the new quantities that may be calculated.

COOP, COHP and DM.

With this addition one can easily calculate the energy resolved LDOS
using either the Green function DM or the spectral function DM.
Together with sisl this enables a very nice addition to the NEGF calculations.