Allow the option of not separating SR and SO in SOC
In a SO+offsite calculation, the use of the fdf line:

    soc-split-sr-so  F

will prevent the program from splitting the lj NL contributions to the
Hamiltonian and energies into SR (ion) and SO parts. The splitting
procedure can be ill-defined. 

There remains a cosmetic issue: Enl is reported as zero, and Eso now
contains the combined former values of Enl and Eso. It can be argued
that a new label "Enl-lj" might be used instead.