Restructured k-point sampling
Now k-points may be user-supplied for further finetuning.
This may be extremely useful in various situations:
1) Detailed analysis on PDOS for spin-orbit calculations
2) Faster convergence with fewer k-points
3) Improved convergence for bias calculations by adding zoom-in
   k-points in the bias-window.
4) And others.

Now the kpoints are associated with a type:
  kpoint_t
this will allow one to read k-points using kgrid.File flag.
The precedence is much like the older implementation:

1) kgrid.MonkhorstPack
2) kgrid.Cutoff
3) kgrid.File

in that order. If none are specified, then the Gamma-point will be used.

This meant that several places the code needed updating.
- Projected DOS k-points
- TranSiesta k-points
- SCF k-points

Now all references to the kpoint_grid module is using kpoint_scf_m module
which uses the type.
This not only reduces lots of duplicated code, but it also allows
one implementation to benefit across all code segments.

Also added a test that reads in the k-points from a file (si_coop_kp_file).