Poperly tagged Enl and Eso when there is no SR-SO split. Extension to onsite.
When the SO contribution is not split (in the 'offsite' flavor with lj projectors,
or now also in the 'onsite' flavor for generality), the Enl energy is tagged 'Enl(+so)'
and the Eso energy is tagged 'Eso(nil)' in the output file.

The CML file is not thus changed (but there is a new parameter 'Split-SR-SO').