Improve and document the handling of the SOC SR-SO split. New ref. outputs
The splitting procedure for the off-site flavor of SOC (involving full
lj projectors) can sometimes be ill-defined, and in those cases the
program relies on a heuristic to compute the two contributions. A
(single) warning is printed (this is a fix from a previous patch).

The use of

    soc-split-sr-so  F

will prevent the program from attempting the splitting (but it still
will be able to detect a possible problem and report an informational
message).

The reference outputs for the 'SOC' tests had been generated by a
non-tracked version during merging, and showed irreproducible
differences with other versions. They have been re-generated.

(+ some easy optimizations in nlefsm_SO_off)