HSX files always contain the 'indxuo' array
The HSX files now output information in 'auxiliary supercell' mode
(former 'gamma' .false.)  even if the supercell is trivially the unit
cell.

Code in Util/COOP has been changed to avoid a spurious check on equivalence
of the two 'gamma's (one indicating whether wavefunction info is complex or
real, and the other the presence of a supercell).