Reimplemented the NaiveAuxiliaryCell option
And further enhanced it by forcing the aux-cell size to
be the largest one for anything but single-shot calculations.

This does not fix *all* cases of error handling since when performing
variable cell calculations the auxiliary cell may again change.

One solution could be if naive aux-cell is used then the aux-cell is
forced to be the same for cell_can_change calculations?